Adding sendVerificationEmail logic to register	321ef8c	Cherry pick
