+++ src/resolvers/users.js
@@ -1,26 +1,6 @@
 const { UserInputError } = require('apollo-server-errors');
 const models = require('../models');
-const { sendEmail } = require('../services/sendEmail');
-
-// TO DO: Edit inivitation email subject, html, and generate email verification link
-const createVerificationEmail = id => {
-  const user = models.User.findByPk(id);
-  if (user == null) {
-    throw new UserInputError('User not found');
-  }
-
-  const hash = user.verificationHash;
-
-  if (hash == null) {
-    // TODO: hash not found?
-  }
-
-  return {
-    from: 'hongyichen@uwblueprint.org',
-    subject: `Verify Email for Building Up`,
-    html: 'Thank you for signing up with <strong>Building Up</strong>. Please join using this link: ',
-  };
-};
+const { createVerificationEmail, sendEmail } = require('../services/sendEmail');
 
 const usersResolvers = {
   Query: {
